fix(first-run): make the Windows CI net hermetic + split the shim under the file-size floor
The new first-run-windows job's first live run caught two things:

- The integration harness was not hermetic on Windows: GitHub's runner
  env carries an UPPERCASE NPM_EXECPATH, and Windows child-env lookup is
  case-insensitive, so the uppercase variant shadowed the test's
  lowercase stub override — the shim silently drove REAL registry npm
  (37-118s per test) and every stub assertion read empty. Verified
  empirically against Windows Node 24 from WSL. runShim now strips every
  case variant before injecting the stub, and every status/log assertion
  carries the child's stdout/stderr so a CI-only failure is diagnosable
  from the log.
- Our own self-guardrail blocked the PR: the rewrite pushed
  packages/create-dxkit/index.js to 518 lines, over the repo's 500-line
  budget. Fixed the finding, not the baseline: pure helpers moved to
  lib.js (index.js re-exports them, so consumers and tests are
  unchanged); lib.js added to the package files list.
