+++ packages/create-dxkit/package.json
@@ -22,6 +22,7 @@
   },
   "files": [
     "index.js",
+    "lib.js",
     "README.md",
     "LICENSE"
   ],
